Делит актуальный код на модули
